Goodbye Delegating, hello WorkflowHandle
Closes #87

`Delegating` was a confusing interface, an artifact of how I backed into
stitching together composite workflow states.

It's replaced by `WorkflowHandle`, a sealed class with two concrete types:

 - `RunningWorkflow`, which includes the workflows current (or bootstrap) state
 - `FinishedWorkflow`, which includes its result

Compositing now looks a lot like it did when we were still hanging live
workflows off of state objects, but everything remains a value type.

We also introduce a convention of reactor companion objects providing
a `run() : RunningWorkflow<...>` methods, for easy bootstrapping.

E.g., this is the state of the `RunGameReactor` while it's
delegating to the `TakeTurnsReactor`:

```
  internal data class Playing(
    val takingTurns: RunningWorkflow<Turn, TakeTurnsEvent, CompletedGame>
  ) : RunGameState() {
    constructor(turn: Turn) : this(TakeTurnsReactor.run(turn))
  }
```

This is how that reactor behaves in that state:

```
    is Playing -> events.select {
      workflows.onWorkflowUpdate(state.takingTurns) {
        when (it) {
          is RunningWorkflow -> EnterState(Playing(it))
          is FinishedWorkflow -> when (it.result.ending) {
            Quitted -> EnterState(MaybeQuitting(it.result))
            else -> EnterState(GameOver(it.result))
          }
        }
      }
    }

```

and here is how it's rendered:

```
      is Playing -> {
        return TakeTurnsRenderer
            .render(state.takingTurns.state, workflows.input(state.takingTurns), workflows)
            .let { MainAndModalScreen(it) }
      }
```

A composite workflow running several others in parallel might look like this:

```
  data class LeftAndRightState(
    val leftWorkflow: RunningWorkflow<LeftState, LeftEvent, LeftResult>,
    val rightWorkflow: RunningWorkflow<RigthState, RightEvent, RightResult>,
  ) : LeftAndRightState()
```

```
    is LeftAndRightState -> events.select {
      workflows.onWorkflowUpdate(state.leftWorfklow) {
        when (it) {
          is RunningWorkflow -> EnterState(state.copy(leftWorkflow = it))
          is FinishedWorkflow -> when (it.result.ending) { ... }
        }
      }

      workflows.onWorkflowUpdate(state.rightWorfklow) {
        when (it) {
          is RunningWorkflow -> EnterState(state.copy(rightWorkflow = it))
          is FinishedWorkflow -> when (it.result.ending) { ... }
        }
      }
    }

```

Also tidies up `ShellReactor` and `ShellState` a bit, to make them more
consistent with the others. This is done with an eye toward #97.